backend/pyproject.toml
[tool.poetry]
name = "quizlaw"
version = "0.1.0"
description = "Legal MCQ Generator with sophisticated UI/UX"
authors = ["QuizLaw Team"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["email"], version = "^2.0.0"}
python-dotenv = "^1.0.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.10.0"
gunicorn = "^21.0.0"
openai = "^1.0.0"
httpx = "^0.24.0"
tenacity = "^8.2.0"
flask-pydantic = "^0.11.0"
cachetools = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.1.0"
ruff = "^0.0.260"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
select = ["E", "F", "I", "N", "B", "SIM", "C90", "UP"]
ignore = []
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true